+++ lib/mitie/text_categorizer.rb
@@ -15,7 +15,7 @@ module Mitie
     end
 
     def categorize(text)
-      tokens = text.is_a?(Array) ? text : Utils.tokenize(text)
+      tokens = text.is_a?(Array) ? text : Mitie.tokenize(text)
       tokens_pointer = Utils.array_to_pointer(tokens)
       text_tag = Fiddle::Pointer.malloc(Fiddle::SIZEOF_VOIDP)
       text_score = Fiddle::Pointer.malloc(Fiddle::SIZEOF_DOUBLE)