+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.1 (unreleased)
 
+- Added `tokenize` method
 - Added support for untokenized text to text categorization
 
 ## 0.2.0 (2022-06-01)